+++ source/bin.ts
@@ -1,5 +1,9 @@
 #!/usr/bin/env node
 import getMAC from './'
-getMAC()
-	.catch(console.error)
-	.then(console.log)
+import { networkInterfaces } from 'os'
+try {
+	console.log(getMAC(process.argv[2]))
+} catch (err) {
+	console.error(networkInterfaces())
+	throw err
+}